Added files from starting Unreal Engine Save
Many files were added in this commit. At present, in Jesse Branch
there now exist the files necessary to commit for the XtremePong
unreal save file. The files combine in Unreal to have a starting
level 0 that is a large plane lit by a directed light source.
The level is still essentially the default. This blank level will
need to be loaded in Unreal, named as startingLevel0, or something
like that. For now, these files exist just on branch jesse.