Made the scoreboard live updating. Also modified defaultLevel0:
At present, the scoreboard class is connected to a blueprint class
in the defaultLevel0 level that displays the player scores
and round number. The score board updates based on an instance
variable frame counter that is updated when the tick function is
called. The next steps will involve connecting the scoreboard to
collision zones, so a ball crossing a goal zone causes the
scoreboard to actually update.