Add paddle collision
Paddles will now properly collide with objects and walls, preventing
them from phasing out of the play area.
Velocity Multiplier was replaced with a Speed Multiplier that sets the
magnitude of the movement vector for each paddle. This value is
accessible in the editor to be changed on the fly.