Added Tests and changed some scoreboard functionality
Five tests were added to the unit testing file for the scoreboard
class. The tests that were added were: incrementation test for
player 2, a test of isWin for player1, an isWin test for player2,
an isWin test for no winners, and a score reset test. More tests
will be added soon ish.