Got scoreboard to manage ball class correctly ish
The scoreboard now spawns a ball that heads towards a player.
During a score, the score is changed and the ball respawns
following the same velocity. I will need to change that soon ish.
I will also have to figure out still how to make the ball explode.
One thing at a time.