Modified some code to make the integration test actually run
The create ball method of AScoreBoard was modified so it just
instantiates the ball class, rather than spawning the actor.
This results in a test that doesn't eat Unreal Engine the
nanosecond it runs. The ball no longer spawns or respawns in the
game, but the test actually runs now. The integration test
tests the scoreboard class' ability to make a ball. It tests to
see if the coordinates are correct. The coordinates that are
correct are <0, 0, 20>. This test fails, since the coordinates are
wrong. It's fine if the test fails. It's at least running now.

The actual test was modified too a bit so as to test for the
ball's location.